fix(GameController): 에외 처리 부분 수정
재귀로 불려야 했던 로직이 잘 불리지 않아 에러처리가 한번만 처리되고 있었고 이를 수정
