+++ src/main/java/javaracingcar/controller/GameController.java
@@ -2,10 +2,10 @@ package javaracingcar.controller;
 
 import javaracingcar.domain.Car;
 import javaracingcar.domain.Game;
+import javaracingcar.view.InputView;
 import javaracingcar.view.OutputView;
 import utils.RandomUtils;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -14,21 +14,29 @@ public class GameController {
     public static final int MIN_RANDOM_NUMBER = 0;
     public static final int MAX_RANDOM_NUMBER = 9;
 
-    public static List<Car> generateCars(List<String> carNames) {
-        List<Car> cars = new ArrayList<>();
-        for (String name : carNames) {
-            cars.add(Car.generateCar(name));
+    public static void start() {
+        try {
+            run();
+        } catch (IllegalArgumentException e) {
+            OutputView.printError(e.getMessage());
+            start();
         }
-        return cars;
     }
 
-    public static void run(List<String> carNames, int trial) throws IllegalArgumentException {
-        Game game = Game.init(carNames, trial);
-        race(game);
+    public static void run() {
+        Game game = initGame();
+        playRace(game);
         OutputView.printWinners(getWinners(game));
     }
 
-    private static void race(Game game) {
+    private static Game initGame() {
+        List<String> carNames = InputController.getCarNames(InputView.receiveCarNamesFromUser());
+        int trial = InputController.getTrial(InputView.receiveTrialFromUser());
+
+        return Game.init(carNames, trial);
+    }
+
+    private static void playRace(Game game) {
         OutputView.printResultTitle();
         while (!game.isEnd()) {
             playOneRound(game);
