+++ src/test/java/javaracingcar/controller/GameControllerTest.java
@@ -1,15 +1,10 @@
 package javaracingcar.controller;
 
-import javaracingcar.domain.Car;
 import javaracingcar.domain.Game;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class GameControllerTest {
     Game game;
@@ -22,16 +17,5 @@ class GameControllerTest {
     }
     // TODO: GameController 테스트를 구현해야 함.
 
-    @Test
-    void generateCars_자동차리스트생성() {
-        List<String> cars = new ArrayList<>();
-        cars.add("CarA");
-        cars.add("CarB");
-        cars.add("CarC");
-        List<Car> generatedCars = GameController.generateCars(Arrays.asList("CarA", "CarB", "CarC"));
-        for (int i = 0; i < cars.size(); i++) {
-            assertEquals(generatedCars.get(i)
-                                      .getName(), cars.get(i));
-        }
-    }
+
 }
