+++ src/test/java/javaracingcar/domain/GameTest.java
@@ -2,6 +2,7 @@ package javaracingcar.domain;
 
 import org.junit.jupiter.api.Test;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -9,6 +10,19 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class GameTest {
+    @Test
+    void generateCars_자동차리스트생성() {
+        List<String> cars = new ArrayList<>();
+        cars.add("CarA");
+        cars.add("CarB");
+        cars.add("CarC");
+        List<Car> generatedCars = Game.generateCars(Arrays.asList("CarA", "CarB", "CarC"));
+        for (int i = 0; i < cars.size(); i++) {
+            assertEquals(generatedCars.get(i)
+                                      .getName(), cars.get(i));
+        }
+    }
+
     @Test
     void init_중복된이름() {
         List<String> carNames = Arrays.asList("a", "b", "c", "a");
